repair: extract repair_reader
