+++ configure.py
@@ -857,6 +857,7 @@ scylla_core = (['database.cc',
                 'repair/hash.cc',
                 'repair/sink_source_for_repair.cc',
                 'repair/row.cc',
+                'repair/reader.cc',
                 'repair/repair.cc',
                 'repair/row_level.cc',
                 'exceptions/exceptions.cc',
